[Metrics UI] Filter out APM nodes from the inventory view (#110300)
* [Metrics UI] Filter out APM nodes from the inventory view

* Update jest snapshots

* Add tests for fs for filtering out APM nodes
