+++ x-pack/plugins/infra/server/routes/snapshot/lib/get_nodes.ts
@@ -10,7 +10,7 @@ import { ESSearchClient } from '../../../lib/metrics/types';
 import { InfraSource } from '../../../lib/sources';
 import { transformRequestToMetricsAPIRequest } from './transform_request_to_metrics_api_request';
 import { queryAllData } from './query_all_data';
-import { transformMetricsApiResponseToSnapshotResponse } from './trasform_metrics_ui_response';
+import { transformMetricsApiResponseToSnapshotResponse } from './transform_metrics_ui_response';
 import { copyMissingMetrics } from './copy_missing_metrics';
 import { LogQueryFields } from '../../../services/log_queries/get_log_query_fields';
